@@ -1,31 +1,35 @@
 {
-  "_from": "mime-db@~1.37.0",
-  "_id": "mime-db@1.37.0",
+  "_args": [
+    [
+      "mime-db@1.38.0",
+      "C:\\Users\\Hasee\\Desktop\\server"
+    ]
+  ],
+  "_from": "mime-db@1.38.0",
+  "_id": "mime-db@1.38.0",
   "_inBundle": false,
-  "_integrity": "sha1-C2oM5v2+lXbiXx8tL96IMNwK0Ng=",
+  "_integrity": "sha1-GiqrFtqesWe0nG5N8tnGjWPY4q0=",
   "_location": "/mime-db",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "version",
     "registry": true,
-    "raw": "mime-db@~1.37.0",
+    "raw": "mime-db@1.38.0",
     "name": "mime-db",
     "escapedName": "mime-db",
-    "rawSpec": "~1.37.0",
+    "rawSpec": "1.38.0",
     "saveSpec": null,
-    "fetchSpec": "~1.37.0"
+    "fetchSpec": "1.38.0"
   },
   "_requiredBy": [
     "/mime-types"
   ],
-  "_resolved": "http://registry.npm.taobao.org/mime-db/download/mime-db-1.37.0.tgz",
-  "_shasum": "0b6a0ce6fdbe9576e25f1f2d2fde8830dc0ad0d8",
-  "_spec": "mime-db@~1.37.0",
-  "_where": "C:\\Users\\Hasee\\Desktop\\study_react\\server\\node_modules\\mime-types",
+  "_resolved": "http://registry.npm.taobao.org/mime-db/download/mime-db-1.38.0.tgz",
+  "_spec": "1.38.0",
+  "_where": "C:\\Users\\Hasee\\Desktop\\server",
   "bugs": {
     "url": "https://github.com/jshttp/mime-db/issues"
   },
-  "bundleDependencies": false,
   "contributors": [
     {
       "name": "Douglas Christopher Wilson",
@@ -42,22 +46,21 @@
       "url": "http://github.com/broofa"
     }
   ],
-  "deprecated": false,
   "description": "Media Type Database",
   "devDependencies": {
-    "bluebird": "3.5.2",
+    "bluebird": "3.5.3",
     "co": "4.6.0",
     "cogent": "1.0.1",
-    "csv-parse": "2.5.0",
-    "eslint": "5.7.0",
+    "csv-parse": "3.2.0",
+    "eslint": "5.13.0",
     "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-import": "2.16.0",
     "eslint-plugin-node": "7.0.1",
     "eslint-plugin-promise": "4.0.1",
     "eslint-plugin-standard": "4.0.0",
     "gnode": "0.1.2",
     "mocha": "5.2.0",
-    "nyc": "13.1.0",
+    "nyc": "13.2.0",
     "raw-body": "2.3.3",
     "stream-to-array": "2.3.0"
   },
@@ -94,7 +97,8 @@
     "test": "mocha --reporter spec --bail --check-leaks test/",
     "test-cov": "nyc --reporter=html --reporter=text npm test",
     "test-travis": "nyc --reporter=text npm test",
-    "update": "npm run fetch && npm run build"
+    "update": "npm run fetch && npm run build",
+    "version": "node scripts/version-history.js && git add HISTORY.md"
   },
-  "version": "1.37.0"
+  "version": "1.38.0"
 }
