@@ -1,32 +1,36 @@
 {
-  "_from": "mime-types@~2.1.18",
-  "_id": "mime-types@2.1.21",
+  "_args": [
+    [
+      "mime-types@2.1.22",
+      "C:\\Users\\Hasee\\Desktop\\server"
+    ]
+  ],
+  "_from": "mime-types@2.1.22",
+  "_id": "mime-types@2.1.22",
   "_inBundle": false,
-  "_integrity": "sha1-KJlaoey3cHQv5q5+WPkYHHRLP5Y=",
+  "_integrity": "sha1-/ms1WhkJJqt2mMmgVWoRGZshmb0=",
   "_location": "/mime-types",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "version",
     "registry": true,
-    "raw": "mime-types@~2.1.18",
+    "raw": "mime-types@2.1.22",
     "name": "mime-types",
     "escapedName": "mime-types",
-    "rawSpec": "~2.1.18",
+    "rawSpec": "2.1.22",
     "saveSpec": null,
-    "fetchSpec": "~2.1.18"
+    "fetchSpec": "2.1.22"
   },
   "_requiredBy": [
     "/accepts",
     "/type-is"
   ],
-  "_resolved": "http://registry.npm.taobao.org/mime-types/download/mime-types-2.1.21.tgz",
-  "_shasum": "28995aa1ecb770742fe6ae7e58f9181c744b3f96",
-  "_spec": "mime-types@~2.1.18",
-  "_where": "C:\\Users\\Hasee\\Desktop\\study_react\\server\\node_modules\\accepts",
+  "_resolved": "http://registry.npm.taobao.org/mime-types/download/mime-types-2.1.22.tgz",
+  "_spec": "2.1.22",
+  "_where": "C:\\Users\\Hasee\\Desktop\\server",
   "bugs": {
     "url": "https://github.com/jshttp/mime-types/issues"
   },
-  "bundleDependencies": false,
   "contributors": [
     {
       "name": "Douglas Christopher Wilson",
@@ -44,19 +48,18 @@
     }
   ],
   "dependencies": {
-    "mime-db": "~1.37.0"
+    "mime-db": "~1.38.0"
   },
-  "deprecated": false,
   "description": "The ultimate javascript content-type utility.",
   "devDependencies": {
-    "eslint": "5.7.0",
+    "eslint": "5.13.0",
     "eslint-config-standard": "12.0.0",
-    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-import": "2.16.0",
     "eslint-plugin-node": "7.0.1",
     "eslint-plugin-promise": "4.0.1",
     "eslint-plugin-standard": "4.0.0",
     "mocha": "5.2.0",
-    "nyc": "13.1.0"
+    "nyc": "13.3.0"
   },
   "engines": {
     "node": ">= 0.6"
@@ -83,5 +86,5 @@
     "test-cov": "nyc --reporter=html --reporter=text npm test",
     "test-travis": "nyc --reporter=text npm test"
   },
-  "version": "2.1.21"
+  "version": "2.1.22"
 }
