@@ -1,9 +1,41 @@
-const React = require('react');
+import React, { Component } from 'react';
 
-const Home = () => {
-  return <div>home</div>
+import Header from '../../components/Header';
+import { connect } from 'react-redux';
+import { getHomeList } from './store/action';
+
+
+
+class Home extends Component {
+  getList() {
+    const { list } = this.props;
+    return list.map(item => <div key={item.id}>{item.title}</div>)
+  }
+  render() {
+    return (
+      <div>
+        <Header />
+        <div>this is {this.props.name}</div>
+        {this.getList()}
+        <button onClick={() => {alert('click')}}>click</button>
+      </div>
+    )
+  }
+
+  componentDidMount() {
+    this.props.getHomeList();
+  }
 }
 
-module.exports = {
-  default: Home
-}
+const mapStateToProps = state => ({
+  list: state.home.newsList,
+  name: state.home.name
+})
+
+const mapDispatchToProps = dispatch => ({
+  getHomeList() {
+    dispatch(getHomeList());
+  }
+})
+
+export default connect(mapStateToProps, mapDispatchToProps)(Home);
