@@ -1,27 +1,18 @@
 const path = require('path');
 const nodeExternals = require('webpack-node-externals');
 
-module.exports = {
+const merge = require('webpack-merge');
+const config = require('./webpack.base.js');
+
+const serverConfig = {
   target: 'node',
   mode: 'development',
-  entry: './src/index.js',
+  entry: './src/server/index.js',
   output: {
     filename: 'bundle.js',
     path: path.resolve(__dirname, 'build')
   },
-  externals: [nodeExternals()],
-  module: {
-    rules: [{
-      test: /\.js?$/,
-      loader: 'babel-loader',
-      exclude: /node_modules/,
-      options: {
-        presets: ['react', 'stage-0',['env', {
-          targets: {
-            browsers: ['last 2 versions']
-          }
-        }]]
-      }
-    }]
-  }
-}
+  externals: [nodeExternals()], // 使require的一些 不打包 比如express不会被webpack打包
+};
+
+module.exports = merge(config, serverConfig);
